requirements.txt
asgiref==3.2.10
bcrypt==3.2.0
cffi==1.14.2
dj-database-url==0.5.0
Django==3.1
django-cors-headers==3.4.0
django-heroku==0.3.1

djangorestframework==3.11.0
djangotoolbox==1.8.0
djongo==1.3.3
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.8.6
pycparser==2.20

pytz==2020.1
six==1.15.0
sqlparse==0.2.4
whitenoise==5.2.0